Some fixes in auth - alerts - mob.  incidents

+++ frontend/src/pages/Profile/index.js
@@ -1,22 +1,22 @@
 import React, { useState, useEffect } from 'react';
+import { Link, useHistory } from 'react-router-dom';
+import { FiPower, FiTrash2, FiSkipBack, FiSkipForward } from 'react-icons/fi';
 import Swal from 'sweetalert2';
 
-import { Link, useHistory } from 'react-router-dom';
-
-import { FiPower, FiTrash2, FiSkipBack, FiSkipForward } from 'react-icons/fi';
 import { Container, Header, Button, PageController } from './styles';
 
 import logo from '../../assets/logo.svg';
-
 import api from '../../services/api';
 
 export default function Profile() {
   const [incidents, setIncidents] = useState([]);
-
   const [page, setPage] = useState(1);
   const [totalPage, setTotalPage] = useState();
 
+  // eslint-disable-next-line camelcase
+  const ong_id = localStorage.getItem('id');
   const ongName = localStorage.getItem('ongName');
+
   const history = useHistory();
 
   function loadIncidents() {
@@ -26,7 +26,8 @@ export default function Profile() {
       })
       .then((response) => {
         setIncidents(response.data.rows);
-        setTotalPage(response.data.count / 4);
+        setTotalPage(Math.floor(Math.ceil(response.data.count / 4)));
+        console.log('total page no laod', totalPage);
       });
   }
 
@@ -36,26 +37,53 @@ export default function Profile() {
   }
 
   async function handleDeleteIncident(id, title) {
-    try {
-      await api.delete(`incidents/delete/${id}`);
-      Swal.fire({
-        title: 'Sucesso',
-        text: `Caso: ${title} deletado da base de dados com sucesso.`,
-        icon: 'success',
-        confirmButtonColor: '#e02041',
-        confirmButtonText: 'Ok!',
-      });
-      setIncidents(incidents.filter((incident) => incident.id !== id));
-    } catch (error) {
-      // trocar por sweet alert
-      alert('Não foi possível deletar - Tente novamente.');
-    }
+    await Swal.fire({
+      title: 'Deletar Caso',
+      text: `Você tem certeza que quer deletar o caso ${title}`,
+      icon: 'warning',
+      showCancelButton: true,
+      cancelButtonColor: '#e02041',
+      confirmButtonColor: '#444',
+      reverseButtons: true,
+      confirmButtonText: 'Sim',
+      cancelButtonText: 'Cancelar',
+    }).then(async (result) => {
+      if (result.value) {
+        try {
+          await api.delete(`incidents/delete/${id}`);
+          console.log('total', totalPage);
+          console.log('atal', page);
+          loadIncidents();
+          console.log('loaded');
+          console.log('VERFIC', totalPage !== page);
+          if (totalPage !== page) {
+            setPage(page - 1);
+          }
+          console.log('total', totalPage);
+          console.log('atal', page);
+          Swal.fire({
+            title: 'Sucesso',
+            text: `Caso ${title} deletado com sucesso.`,
+            icon: 'success',
+            confirmButtonColor: '#e02041',
+            confirmButtonText: 'Okay',
+          });
+        } catch (error) {
+          // trocar por sweet alert
+          Swal.fire({
+            title: 'Que pena.',
+            text: `Caso: ${title} não pode ser deletado. Tente novamente.`,
+            icon: 'error',
+            confirmButtonColor: '#e02041',
+            confirmButtonText: 'Okay',
+          });
+        }
+      }
+    });
   }
 
   useEffect(() => {
     loadIncidents();
-    console.log('PAGINA ATUAL', page);
-    console.log('TOTAL E PAGINAS', totalPage);
   }, [page]);
 
   return (
@@ -71,9 +99,7 @@ export default function Profile() {
           <FiPower size={48} color="#e02048" />
         </Button>
       </Header>
-
       <h1> Casos Cadastrados </h1>
-
       <ul>
         {incidents.map((incident) => (
           <li key={incident.id}>
@@ -114,7 +140,6 @@ export default function Profile() {
         <p>{page}</p>
         <Button
           onClick={() => {
-            // console.log('QUANTIDADE DE PAGINAS', pagesQuantity);
             if (page !== totalPage) {
               setPage(page + 1);
             }
